feat: 노선 이름을 저장하는 기능 구현

+++ src/main/java/subway/service/LineService.java
@@ -3,6 +3,8 @@ package subway.service;
 import java.util.List;
 import subway.domain.Line;
 import subway.domain.LineRepository;
+import subway.global.exception.CustomException;
+import subway.global.exception.ErrorMessage;
 
 public class LineService implements SubwayService {
     private final LineRepository lineRepository;
@@ -20,11 +22,22 @@ public class LineService implements SubwayService {
 
     @Override
     public void add(String line) {
+        if (lineRepository.findByName(line).isPresent()) {
+            throw CustomException.from(ErrorMessage.LINE_DUPLICATED_ERROR);
+        }
         lineRepository.addLine(new Line(line));
     }
 
     @Override
     public void delete(String line) {
+        if (lineRepository.findByName(line).isEmpty()) {
+            throw CustomException.from(ErrorMessage.LINE_NOT_FOUND_ERROR);
+        }
         lineRepository.deleteLineByName(line);
+    }
+
+    @Override
+    public List<String> getAll() {
+        return null;
     }
 }
